update edit view
